🚀 Update Main Layout
재영님께서 멋드러지게 작업했습니다.
